Fixed Automated_Scanning_V1 - should work
Currently having slight trouble opening in an actual excel file, not a
matlab generated excel file. However judging by the errors given I
believe it is a problem with the matlab-excel compadibility as the exact
same code (line-for-line) did not have this problem yesterday. This has
not been tested yet, but I believe it will work on another computer.